refactor(player): extract PlayerScreen components into ui/player (#94)
