+++ native/app/src/main/java/com/localstream/app/ui/player/NextEpisodeOverlay.kt
@@ -1,0 +1,90 @@
+package com.localstream.app.ui.player
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableIntStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.localstream.app.domain.model.VideoItem
+import com.localstream.app.ui.theme.Red600
+import com.localstream.app.ui.theme.White
+import com.localstream.app.ui.theme.Zinc800
+import com.localstream.app.ui.theme.Zinc900
+import kotlinx.coroutines.delay
+
+@Composable
+fun NextEpisodeOverlay(
+    nextVideo: VideoItem,
+    onPlayNext: () -> Unit,
+    onDismiss: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    var countdown by remember { mutableIntStateOf(5) }
+
+    LaunchedEffect(nextVideo) {
+        countdown = 5
+        while (countdown > 0) {
+            delay(1000L)
+            countdown--
+        }
+        onPlayNext()
+    }
+
+    Card(
+        shape = RoundedCornerShape(16.dp),
+        colors = CardDefaults.cardColors(containerColor = Zinc900),
+        modifier = modifier.fillMaxWidth(),
+    ) {
+        Column(modifier = Modifier.padding(16.dp)) {
+            Text("Épisode suivant dans $countdown s", color = Red600, fontSize = 12.sp, fontWeight = FontWeight.Bold)
+            Spacer(modifier = Modifier.height(4.dp))
+            Text(
+                text = nextVideo.cleanTitle ?: nextVideo.name,
+                color = White,
+                fontSize = 16.sp,
+                fontWeight = FontWeight.Bold,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis,
+            )
+            Spacer(modifier = Modifier.height(12.dp))
+            Row(
+                horizontalArrangement = Arrangement.End,
+                modifier = Modifier.fillMaxWidth(),
+            ) {
+                Button(
+                    onClick = onDismiss,
+                    colors = ButtonDefaults.buttonColors(containerColor = Zinc800),
+                ) {
+                    Text("Fermer", color = White)
+                }
+                Spacer(modifier = Modifier.width(8.dp))
+                Button(
+                    onClick = onPlayNext,
+                    colors = ButtonDefaults.buttonColors(containerColor = Red600),
+                ) {
+                    Text("Lire maintenant", color = White)
+                }
+            }
+        }
+    }
+}
