--- native/app/src/main/java/com/localstream/app/ui/screens/PlayerScreen.kt
@@ -1,1538 +1,0 @@
-@file:OptIn(
-    androidx.compose.material3.ExperimentalMaterial3Api::class,
-    androidx.media3.common.util.UnstableApi::class,
-)
-
-package com.localstream.app.ui.screens
-
-import android.app.Activity
-import android.app.PictureInPictureParams
-import android.content.Context
-import android.content.Intent
-import android.content.pm.ActivityInfo
-import android.media.AudioManager
-import android.net.Uri
-import android.os.Build
-import android.util.Rational
-import android.view.ViewGroup
-import android.view.WindowManager
-import androidx.activity.compose.BackHandler
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.gestures.awaitEachGesture
-import androidx.compose.foundation.gestures.awaitFirstDown
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.AspectRatio
-import androidx.compose.material.icons.filled.Brightness6
-import androidx.compose.material.icons.filled.ErrorOutline
-import androidx.compose.material.icons.filled.Forward10
-import androidx.compose.material.icons.filled.Lock
-import androidx.compose.material.icons.filled.LockOpen
-import androidx.compose.material.icons.filled.Pause
-import androidx.compose.material.icons.filled.PictureInPicture
-import androidx.compose.material.icons.filled.PlayArrow
-import androidx.compose.material.icons.filled.Replay10
-import androidx.compose.material.icons.filled.SkipNext
-import androidx.compose.material.icons.filled.Speed
-import androidx.compose.material.icons.filled.Subtitles
-import androidx.compose.material.icons.filled.VolumeUp
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.RadioButton
-import androidx.compose.material3.RadioButtonDefaults
-import androidx.compose.material3.Slider
-import androidx.compose.material3.SliderDefaults
-import androidx.compose.material3.Text
-import androidx.compose.material3.rememberModalBottomSheetState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableLongStateOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.input.pointer.PointerInputScope
-import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.compose.ui.viewinterop.AndroidView
-import androidx.core.view.WindowCompat
-import androidx.core.view.WindowInsetsCompat
-import androidx.core.view.WindowInsetsControllerCompat
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleEventObserver
-import androidx.lifecycle.compose.LocalLifecycleOwner
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.media3.common.AudioAttributes
-import androidx.media3.common.C
-import androidx.media3.common.MediaItem
-import androidx.media3.common.MimeTypes
-import androidx.media3.common.PlaybackException
-import androidx.media3.common.Player
-import androidx.media3.common.TrackSelectionOverride
-import androidx.media3.common.Tracks
-import androidx.media3.exoplayer.ExoPlayer
-import androidx.media3.ui.AspectRatioFrameLayout
-import androidx.media3.ui.PlayerView
-import android.annotation.SuppressLint
-import android.webkit.JavascriptInterface
-import android.webkit.WebChromeClient
-import android.webkit.WebView
-import android.webkit.WebViewClient
-import com.localstream.app.domain.YoutubeUtils
-import com.localstream.app.LocalStreamApplication
-import com.localstream.app.domain.model.VideoItem
-import com.localstream.app.ui.player.AspectRatioMode
-import com.localstream.app.ui.player.AudioTrackUiState
-import com.localstream.app.ui.player.FeedbackType
-import com.localstream.app.ui.player.GestureFeedback
-import com.localstream.app.ui.player.PlayerViewModel
-import com.localstream.app.ui.player.SubtitleTrackUiState
-import com.localstream.app.ui.theme.Black
-import com.localstream.app.ui.theme.Red600
-import com.localstream.app.ui.theme.White
-import com.localstream.app.ui.theme.Zinc800
-import com.localstream.app.ui.theme.Zinc900
-import java.io.File
-import java.util.Locale
-import kotlin.math.abs
-import kotlin.math.roundToInt
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.delay
-
-private const val CONTROLS_TIMEOUT_MS = 3500L
-private const val FEEDBACK_TIMEOUT_MS = 1500L
-// A full-height vertical swipe spans 100% of the volume range (lower = more sensitive).
-private const val VOLUME_GESTURE_PERCENT = 100f
-private const val DRAG_SEEK_THROTTLE_MS = 120L
-// A full-height vertical swipe spans the whole 5%-100% brightness range (log scale).
-private val BRIGHTNESS_GESTURE_LOG_RANGE: Float =
-    Math.log((PlayerViewModel.MAX_BRIGHTNESS / PlayerViewModel.MIN_BRIGHTNESS).toDouble()).toFloat()
-
-
-@Suppress("LongMethod", "CyclomaticComplexMethod", "TooManyFunctions")
-@Composable
-fun PlayerScreen(
-    videoName: String,
-    onBack: () -> Unit,
-    viewModel: PlayerViewModel = run {
-        val container = (LocalContext.current.applicationContext as LocalStreamApplication).container
-        viewModel(factory = PlayerViewModel.factory(videoName, container))
-    },
-) {
-    val context = LocalContext.current
-    val activity = context as? Activity
-    val uiState by viewModel.uiState.collectAsStateWithLifecycle()
-
-    val youtubeId = remember(uiState.currentVideo) {
-        val video = uiState.currentVideo
-        if (video != null) {
-            YoutubeUtils.extractVideoId(video.url)
-                ?: YoutubeUtils.extractVideoId(video.name)
-                ?: YoutubeUtils.extractVideoId(video.path)
-        } else null
-    }
-    val audioManager = remember(context) {
-        context.getSystemService(Context.AUDIO_SERVICE) as? AudioManager
-    }
-    val lifecycleOwner = LocalLifecycleOwner.current
-
-    // Immersive Landscape mode
-    DisposableEffect(activity) {
-        val window = activity?.window
-        activity?.requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_SENSOR_LANDSCAPE
-        if (window != null) {
-            val insetsController = WindowCompat.getInsetsController(window, window.decorView)
-            insetsController.systemBarsBehavior =
-                WindowInsetsControllerCompat.BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE
-            insetsController.hide(WindowInsetsCompat.Type.systemBars())
-        }
-        onDispose {
-            activity?.requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_UNSPECIFIED
-            if (window != null) {
-                val insetsController = WindowCompat.getInsetsController(window, window.decorView)
-                insetsController.show(WindowInsetsCompat.Type.systemBars())
-            }
-        }
-    }
-
-    // Keep screen turned on for the whole time the player is on screen.
-    // NOTE: this used to be tied to `uiState.isPlaying`, which caused the flag to be
-    // removed and re-added on every buffering stall (isPlaying flips to false while
-    // buffering). If the system's screen-timeout fired during one of those brief gaps,
-    // the screen would go to sleep and never wake back up on its own (see issue #80).
-    // Keeping it on for the entire player session removes that race entirely.
-    DisposableEffect(activity) {
-        val window = activity?.window
-        window?.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
-        onDispose {
-            window?.clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
-        }
-    }
-
-    var isVolumeInitialized by remember { mutableStateOf(false) }
-
-    // Init volume from system AudioManager
-    LaunchedEffect(Unit) {
-        audioManager?.let { am ->
-            val maxVol = am.getStreamMaxVolume(AudioManager.STREAM_MUSIC)
-            if (maxVol > 0) {
-                val curVol = am.getStreamVolume(AudioManager.STREAM_MUSIC)
-                val curPct = ((curVol.toFloat() / maxVol) * 100f).coerceIn(0f, 100f)
-                viewModel.setInitialVolumePercent(curPct)
-            }
-        }
-        isVolumeInitialized = true
-    }
-
-    // Sync volume to AudioManager
-    LaunchedEffect(uiState.volumePercent, isVolumeInitialized) {
-        if (!isVolumeInitialized) return@LaunchedEffect
-        audioManager?.let { am ->
-            val maxVol = am.getStreamMaxVolume(AudioManager.STREAM_MUSIC)
-            if (maxVol > 0) {
-                val targetVol = ((uiState.volumePercent / 100f) * maxVol).roundToInt().coerceIn(0, maxVol)
-                am.setStreamVolume(AudioManager.STREAM_MUSIC, targetVol, 0)
-            }
-        }
-    }
-
-    // External player launcher if selected in settings
-    LaunchedEffect(uiState.playerMode, uiState.currentVideo) {
-        val video = uiState.currentVideo
-        if (uiState.playerMode == "external" && video != null) {
-            launchExternalPlayer(context, video, uiState.selectedExternalPlayer)
-            onBack()
-        }
-    }
-
-    var showTracksSheet by remember { mutableStateOf(false) }
-
-    // SAF picker launcher for external .srt subtitle files
-    val subtitleFilePicker = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.GetContent()
-    ) { uri: Uri? ->
-        uri?.let {
-            val fileName = it.lastPathSegment?.substringAfterLast('/') ?: "Sous-titre"
-            viewModel.addExternalSubtitle(fileName, it.toString())
-        }
-    }
-
-    // Controls overlay auto-hide delay
-    LaunchedEffect(uiState.isControlsVisible, uiState.isPlaying, uiState.isLocked) {
-        if (uiState.isControlsVisible && uiState.isPlaying && !uiState.isLocked) {
-            delay(CONTROLS_TIMEOUT_MS)
-            viewModel.setControlsVisible(false)
-        }
-    }
-
-    // Clear gesture feedback card
-    LaunchedEffect(uiState.gestureFeedback) {
-        if (uiState.gestureFeedback != null) {
-            delay(FEEDBACK_TIMEOUT_MS)
-            viewModel.clearGestureFeedback()
-        }
-    }
-
-    // Window brightness control
-    LaunchedEffect(uiState.brightnessPercent) {
-        val brightness = uiState.brightnessPercent
-        if (brightness >= 0f) {
-            activity?.window?.attributes = activity?.window?.attributes?.apply {
-                screenBrightness = brightness
-            }
-        }
-    }
-    DisposableEffect(activity) {
-        onDispose {
-            activity?.window?.attributes = activity?.window?.attributes?.apply {
-                screenBrightness = WindowManager.LayoutParams.BRIGHTNESS_OVERRIDE_NONE
-            }
-        }
-    }
-
-    // ExoPlayer creation & AudioAttributes
-    val exoPlayer = remember(context) {
-        val audioAttributes = AudioAttributes.Builder()
-            .setUsage(C.USAGE_MEDIA)
-            .setContentType(C.AUDIO_CONTENT_TYPE_MOVIE)
-            .build()
-        ExoPlayer.Builder(context)
-            .setAudioAttributes(audioAttributes, true)
-            .build()
-    }
-
-    // Pause on background lifecycle event, but NOT when the ON_PAUSE is caused by
-    // entering Picture-in-Picture mode (that would freeze the video the instant the
-    // user taps the PiP button, defeating the whole point of the feature).
-    DisposableEffect(lifecycleOwner, exoPlayer) {
-        val observer = LifecycleEventObserver { _, event ->
-            val isEnteringPip = Build.VERSION.SDK_INT >= Build.VERSION_CODES.O &&
-                activity?.isInPictureInPictureMode == true
-            if (event == Lifecycle.Event.ON_PAUSE && !isEnteringPip) {
-                exoPlayer.pause()
-            }
-        }
-        lifecycleOwner.lifecycle.addObserver(observer)
-        onDispose {
-            lifecycleOwner.lifecycle.removeObserver(observer)
-        }
-    }
-
-    var isPlayerReady by remember { mutableStateOf(false) }
-
-    // ExoPlayer event listeners
-    DisposableEffect(exoPlayer) {
-        val listener = object : Player.Listener {
-            override fun onPlaybackStateChanged(state: Int) {
-                when (state) {
-                    Player.STATE_BUFFERING -> viewModel.setBuffering(true)
-                    Player.STATE_READY -> {
-                        viewModel.setBuffering(false)
-                        viewModel.setErrorMessage(null)
-                        isPlayerReady = true
-                        val dur = exoPlayer.duration.coerceAtLeast(0L)
-                        if (dur > 0L) {
-                            viewModel.onPositionChanged(exoPlayer.currentPosition.coerceAtLeast(0L), dur)
-                        }
-                    }
-                    Player.STATE_ENDED -> {
-                        viewModel.setBuffering(false)
-                        viewModel.onVideoEnded()
-                    }
-                    else -> viewModel.setBuffering(false)
-                }
-            }
-
-            override fun onPlayerError(error: PlaybackException) {
-                viewModel.setBuffering(false)
-                viewModel.setErrorMessage(error.localizedMessage ?: "Erreur de lecture de la vidéo")
-            }
-
-            override fun onIsPlayingChanged(isPlaying: Boolean) {
-                viewModel.onPlayingStateChanged(isPlaying)
-            }
-
-            override fun onTracksChanged(tracks: Tracks) {
-                val audioList = mutableListOf<AudioTrackUiState>()
-                val subList = mutableListOf<SubtitleTrackUiState>()
-
-                for (group in tracks.groups) {
-                    val trackType = group.type
-                    for (i in 0 until group.length) {
-                        val format = group.getTrackFormat(i)
-                        val id = format.id ?: "$trackType-$i"
-                        val label = format.label ?: format.language ?: "Piste ${i + 1}"
-                        val isSelected = group.isTrackSelected(i)
-
-                        if (trackType == C.TRACK_TYPE_AUDIO) {
-                            audioList.add(
-                                AudioTrackUiState(
-                                    id = id,
-                                    label = label,
-                                    language = format.language,
-                                    isSelected = isSelected,
-                                )
-                            )
-                        } else if (trackType == C.TRACK_TYPE_TEXT) {
-                            subList.add(
-                                SubtitleTrackUiState(
-                                    id = id,
-                                    label = label,
-                                    language = format.language,
-                                    isSelected = isSelected,
-                                )
-                            )
-                        }
-                    }
-                }
-                viewModel.updateTracks(audioList, subList)
-            }
-        }
-        exoPlayer.addListener(listener)
-        onDispose {
-            exoPlayer.removeListener(listener)
-            exoPlayer.stop()
-            exoPlayer.clearMediaItems()
-            exoPlayer.release()
-        }
-    }
-
-    // Apply track selections to ExoPlayer
-    LaunchedEffect(uiState.selectedAudioTrackId, uiState.selectedSubtitleTrackId, exoPlayer.currentTracks) {
-        val tracks = exoPlayer.currentTracks
-        val builder = exoPlayer.trackSelectionParameters.buildUpon()
-
-        // Audio Track
-        val selAudioId = uiState.selectedAudioTrackId
-        if (selAudioId != null) {
-            for (group in tracks.groups) {
-                if (group.type == C.TRACK_TYPE_AUDIO) {
-                    for (i in 0 until group.length) {
-                        val format = group.getTrackFormat(i)
-                        val id = format.id ?: "${group.type}-$i"
-                        if (id == selAudioId) {
-                            builder.setOverrideForType(TrackSelectionOverride(group.mediaTrackGroup, i))
-                            break
-                        }
-                    }
-                }
-            }
-        }
-
-        // Subtitle Track
-        val selSubId = uiState.selectedSubtitleTrackId
-        if (selSubId == null) {
-            builder.setTrackTypeDisabled(C.TRACK_TYPE_TEXT, true)
-        } else {
-            builder.setTrackTypeDisabled(C.TRACK_TYPE_TEXT, false)
-            for (group in tracks.groups) {
-                if (group.type == C.TRACK_TYPE_TEXT) {
-                    for (i in 0 until group.length) {
-                        val format = group.getTrackFormat(i)
-                        val id = format.id ?: "${group.type}-$i"
-                        if (id == selSubId) {
-                            builder.setOverrideForType(TrackSelectionOverride(group.mediaTrackGroup, i))
-                            break
-                        }
-                    }
-                }
-            }
-        }
-
-        exoPlayer.trackSelectionParameters = builder.build()
-    }
-
-    // Handle imported external subtitle file (.srt)
-    LaunchedEffect(uiState.subtitleTracks) {
-        val externalTrack = uiState.subtitleTracks.firstOrNull { it.isExternal && it.isSelected && !it.uriString.isNullOrEmpty() }
-        if (externalTrack != null) {
-            val currentVideo = uiState.currentVideo ?: return@LaunchedEffect
-            val uri = extractUri(currentVideo) ?: return@LaunchedEffect
-
-            val subUri = Uri.parse(externalTrack.uriString)
-            val subConfig = MediaItem.SubtitleConfiguration.Builder(subUri)
-                .setMimeType(MimeTypes.APPLICATION_SUBRIP)
-                .setLanguage("fr")
-                .setSelectionFlags(C.SELECTION_FLAG_DEFAULT)
-                .build()
-
-            val curPos = exoPlayer.currentPosition
-            val mediaItem = MediaItem.Builder()
-                .setUri(uri)
-                .setSubtitleConfigurations(listOf(subConfig))
-                .build()
-
-            exoPlayer.setMediaItem(mediaItem, curPos)
-            exoPlayer.prepare()
-            exoPlayer.playWhenReady = true
-        }
-    }
-
-    // Load media source into ExoPlayer
-    LaunchedEffect(uiState.currentVideo, youtubeId) {
-        if (youtubeId != null) return@LaunchedEffect
-        val video = uiState.currentVideo ?: return@LaunchedEffect
-        val uri = extractUri(video) ?: return@LaunchedEffect
-
-        exoPlayer.stop()
-       exoPlayer.clearMediaItems()
-        isPlayerReady = false
-        val mediaItem = MediaItem.fromUri(uri)
-        val startPos = uiState.initialPositionMs
-        if (startPos > 0L) {
-            exoPlayer.setMediaItem(mediaItem, startPos)
-        } else {
-            exoPlayer.setMediaItem(mediaItem)
-        }
-        exoPlayer.prepare()
-        exoPlayer.playWhenReady = true
-    }
-
-    // Continuously update position flow
-    LaunchedEffect(exoPlayer) {
-        while (true) {
-            if (isPlayerReady) {
-                viewModel.onPositionChanged(
-                    positionMs = exoPlayer.currentPosition.coerceAtLeast(0L),
-                    durationMs = exoPlayer.duration.coerceAtLeast(0L),
-                )
-            }
-            delay(500L)
-        }
-    }
-
-    // Apply playback speed changes
-    LaunchedEffect(uiState.playbackSpeed) {
-        exoPlayer.setPlaybackSpeed(uiState.playbackSpeed)
-    }
-
-    BackHandler {
-        onBack()
-    }
-
-    // Tracks the running target position of an in-progress horizontal swipe so that
-    // exoPlayer.seekTo() calls can be throttled (see onHorizontalDrag below) without
-    // losing track of the total accumulated drag distance.
-    var pendingSeekTargetMs by remember { mutableStateOf<Long?>(null) }
-    var lastRealSeekAtMs by remember { mutableStateOf(0L) }
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Black)
-            .pointerInput(Unit) {
-                detectPlayerGestures(
-                    PlayerGestureCallbacks(
-                        onTap = {
-                            viewModel.toggleControlsVisibility()
-                        },
-                        onDoubleTapLeft = {
-                            if (!uiState.isLocked) {
-                                viewModel.seekBy(-10000L)
-                                if (youtubeId == null) {
-                                    exoPlayer.seekTo((exoPlayer.currentPosition - 10000L).coerceAtLeast(0L))
-                                }
-                            }
-                        },
-                        onDoubleTapRight = {
-                            if (!uiState.isLocked) {
-                                viewModel.seekBy(10000L)
-                                if (youtubeId == null) {
-                                    exoPlayer.seekTo((exoPlayer.currentPosition + 10000L).coerceAtMost(exoPlayer.duration))
-                                }
-                            }
-                        },
-                        onVerticalDragLeft = { deltaRatio ->
-                            if (!uiState.isLocked) {
-                                viewModel.scaleBrightness(
-                                    Math.exp(deltaRatio * BRIGHTNESS_GESTURE_LOG_RANGE.toDouble()).toFloat(),
-                                )
-                            }
-                        },
-                        onVerticalDragRight = { deltaRatio ->
-                            if (!uiState.isLocked) {
-                                viewModel.setVolumePercent(
-                                    uiState.volumePercent + deltaRatio * VOLUME_GESTURE_PERCENT,
-                                )
-                            }
-                        },
-                        onHorizontalDrag = { ratio ->
-                            if (!uiState.isLocked) {
-                                val dur = if (youtubeId != null) uiState.durationMs else exoPlayer.duration.coerceAtLeast(1L)
-                                // Base the running target off our own accumulator rather than
-                                // exoPlayer.currentPosition, since that position only updates
-                                // when we actually call seekTo() below (which we now throttle).
-                                val basePos = pendingSeekTargetMs
-                                    ?: (if (youtubeId != null) uiState.positionMs else exoPlayer.currentPosition)
-                                val seekOffset = (ratio * 60000L).toLong()
-                                val targetPos = (basePos + seekOffset).coerceIn(0L, dur)
-                                pendingSeekTargetMs = targetPos
-                                viewModel.seekBy(seekOffset)
-                                if (youtubeId == null) {
-                                    // Real seeks are expensive (keyframe lookup + decode), so we
-                                    // only send one every DRAG_SEEK_THROTTLE_MS instead of on every
-                                    // pointer-move delta, which used to cause visible stutter.
-                                    val now = System.currentTimeMillis()
-                                    if (now - lastRealSeekAtMs >= DRAG_SEEK_THROTTLE_MS) {
-                                        lastRealSeekAtMs = now
-                                        exoPlayer.seekTo(targetPos)
-                                    }
-                                }
-                            }
-                        },
-                        onDragEnd = {
-                            // Always land on the exact released position, even if the last
-                            // move event was throttled away.
-                            val target = pendingSeekTargetMs
-                            if (target != null && youtubeId == null) {
-                                exoPlayer.seekTo(target)
-                            }
-                            pendingSeekTargetMs = null
-                        },
-                    )
-                )
-            },
-    ) {
-        // ExoPlayer AndroidView
-        AndroidView(
-            factory = { ctx ->
-                PlayerView(ctx).apply {
-                    player = exoPlayer
-                    useController = false
-                    layoutParams = ViewGroup.LayoutParams(
-                        ViewGroup.LayoutParams.MATCH_PARENT,
-                        ViewGroup.LayoutParams.MATCH_PARENT,
-                    )
-                }
-            },
-            update = { playerView ->
-                playerView.resizeMode = when (uiState.aspectRatioMode) {
-                    AspectRatioMode.FIT -> AspectRatioFrameLayout.RESIZE_MODE_FIT
-                    AspectRatioMode.FILL -> AspectRatioFrameLayout.RESIZE_MODE_FILL
-                    AspectRatioMode.ZOOM -> AspectRatioFrameLayout.RESIZE_MODE_ZOOM
-                    AspectRatioMode.RATIO_16_9 -> AspectRatioFrameLayout.RESIZE_MODE_FIXED_WIDTH
-                    AspectRatioMode.RATIO_4_3 -> AspectRatioFrameLayout.RESIZE_MODE_FIXED_HEIGHT
-                }
-            },
-            modifier = Modifier.fillMaxSize(),
-        )
-
-        // Buffering Indicator Overlay
-        if (uiState.isBuffering && uiState.errorMessage == null) {
-            CircularProgressIndicator(
-                color = Red600,
-                modifier = Modifier
-                    .size(56.dp)
-                    .align(Alignment.Center),
-            )
-        }
-
-        // Error Message Overlay
-        uiState.errorMessage?.let { errorMsg ->
-            Card(
-                shape = RoundedCornerShape(16.dp),
-                colors = CardDefaults.cardColors(containerColor = Zinc900.copy(alpha = 0.95f)),
-                modifier = Modifier
-                    .align(Alignment.Center)
-                    .padding(24.dp),
-            ) {
-                Column(
-                    modifier = Modifier.padding(24.dp),
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                ) {
-                    Icon(
-                        imageVector = Icons.Filled.ErrorOutline,
-                        contentDescription = null,
-                        tint = Red600,
-                        modifier = Modifier.size(48.dp),
-                    )
-                    Spacer(modifier = Modifier.height(12.dp))
-                    Text(
-                        text = "Erreur de lecture",
-                        color = White,
-                        fontSize = 18.sp,
-                        fontWeight = FontWeight.Bold,
-                    )
-                    Spacer(modifier = Modifier.height(8.dp))
-                    Text(
-                        text = errorMsg,
-                        color = Color.LightGray,
-                        fontSize = 14.sp,
-                    )
-                    Spacer(modifier = Modifier.height(16.dp))
-                    Button(
-                        onClick = {
-                            viewModel.retryPlayback()
-                            exoPlayer.prepare()
-                            exoPlayer.play()
-                        },
-                        colors = ButtonDefaults.buttonColors(containerColor = Red600),
-                    ) {
-                        Text("Réessayer", color = White)
-                    }
-                }
-            }
-        }
-
-        // Gesture feedback overlay (Volume / Brightness / Seek)
-        uiState.gestureFeedback?.let { feedback ->
-            GestureFeedbackCard(
-                feedback = feedback,
-                modifier = Modifier.align(Alignment.Center),
-            )
-        }
-
-        // Video controls overlay
-        AnimatedVisibility(
-            visible = uiState.isControlsVisible && !uiState.isLocked,
-            enter = fadeIn(),
-            exit = fadeOut(),
-            modifier = Modifier.fillMaxSize(),
-        ) {
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .background(Color.Black.copy(alpha = 0.6f)),
-            ) {
-                // Top Bar
-                TopPlayerBar(
-                    title = uiState.currentVideo?.cleanTitle ?: uiState.currentVideo?.name ?: videoName,
-                    aspectRatioMode = uiState.aspectRatioMode,
-                    playbackSpeed = uiState.playbackSpeed,
-                    onBack = onBack,
-                    onOpenTracks = { showTracksSheet = true },
-                    onCycleAspect = viewModel::cycleAspectRatio,
-                    onCycleSpeed = viewModel::cyclePlaybackSpeed,
-                    modifier = Modifier.align(Alignment.TopCenter),
-                )
-
-                // Center controls (Rewind, Play/Pause, Fast-Forward, Next Episode)
-                CenterPlayerControls(
-                    isPlaying = uiState.isPlaying,
-                    hasNextVideo = uiState.nextVideo != null,
-                    onTogglePlay = {
-                        if (exoPlayer.isPlaying) {
-                            exoPlayer.pause()
-                        } else {
-                            exoPlayer.play()
-                        }
-                    },
-                    onRewind = {
-                        viewModel.seekBy(-10000L)
-                        exoPlayer.seekTo((exoPlayer.currentPosition - 10000L).coerceAtLeast(0L))
-                    },
-                    onForward = {
-                        viewModel.seekBy(10000L)
-                        exoPlayer.seekTo((exoPlayer.currentPosition + 10000L).coerceAtMost(exoPlayer.duration))
-                    },
-                    onNextVideo = viewModel::playNextVideo,
-                    modifier = Modifier.align(Alignment.Center),
-                )
-
-                // Bottom Bar (Slider, Time, PiP, Lock)
-                BottomPlayerBar(
-                    positionMs = uiState.positionMs,
-                    durationMs = uiState.durationMs,
-                    isLocked = uiState.isLocked,
-                    onSeek = { newPos ->
-                        viewModel.onPositionChanged(newPos, uiState.durationMs)
-                        exoPlayer.seekTo(newPos)
-                    },
-                    onToggleLock = viewModel::toggleLock,
-                    onEnterPip = {
-                        enterPipMode(activity)
-                    },
-                    modifier = Modifier.align(Alignment.BottomCenter),
-                )
-            }
-        }
-
-        // Unlock button when controls are locked
-        if (uiState.isLocked) {
-            IconButton(
-                onClick = viewModel::toggleLock,
-                modifier = Modifier
-                    .align(Alignment.TopEnd)
-                    .padding(24.dp)
-                    .background(Zinc900.copy(alpha = 0.8f), CircleShape),
-            ) {
-                Icon(Icons.Filled.Lock, contentDescription = "Déverrouiller", tint = Red600)
-            }
-        }
-
-        // Next Episode End-of-video overlay with auto countdown
-        if (uiState.isEnded && uiState.nextVideo != null) {
-            NextEpisodeOverlay(
-                nextVideo = uiState.nextVideo!!,
-                onPlayNext = viewModel::playNextVideo,
-                onDismiss = onBack,
-                modifier = Modifier
-                    .align(Alignment.BottomCenter)
-                    .padding(24.dp),
-            )
-        }
-    }
-
-    // Audio & Subtitle Tracks Bottom Sheet
-    if (showTracksSheet) {
-        TracksSelectionSheet(
-            audioTracks = uiState.audioTracks,
-            subtitleTracks = uiState.subtitleTracks,
-            onSelectAudio = viewModel::selectAudioTrack,
-            onSelectSubtitle = viewModel::selectSubtitleTrack,
-            onPickSubtitleFile = {
-                subtitleFilePicker.launch("application/x-subrip")
-            },
-            onDismiss = { showTracksSheet = false },
-        )
-    }
-}
-
-private data class PlayerGestureCallbacks(
-    val onTap: () -> Unit,
-    val onDoubleTapLeft: () -> Unit,
-    val onDoubleTapRight: () -> Unit,
-    val onVerticalDragLeft: (Float) -> Unit,
-    val onVerticalDragRight: (Float) -> Unit,
-    val onHorizontalDrag: (Float) -> Unit,
-    val onDragEnd: () -> Unit = {},
-)
-
-private suspend fun PointerInputScope.detectPlayerGestures(
-    callbacks: PlayerGestureCallbacks,
-) {
-    var lastTapTime = 0L
-    var lastTapX = 0f
-
-    awaitEachGesture {
-        val down = awaitFirstDown(requireUnconsumed = false)
-        val startX = down.position.x
-        val touchSlop = viewConfiguration.touchSlop
-        val pointerId = down.id
-        var isDrag = false
-        var dragMode = 0
-        var pointerActive = true
-        var lastX = startX
-        var lastY = down.position.y
-
-        while (pointerActive) {
-            val event = awaitPointerEvent()
-            val change = event.changes.firstOrNull { it.id == pointerId }
-            if (change == null || !change.pressed) {
-                pointerActive = false
-                if (isDrag) {
-                    callbacks.onDragEnd()
-                } else {
-                    lastTapTime = processTapRelease(startX, size.width.toFloat(), lastTapTime, lastTapX, callbacks)
-                    lastTapX = startX
-                }
-            } else {
-                val totalDx = change.position.x - startX
-                val totalDy = change.position.y - down.position.y
-                if (!isDrag && (abs(totalDx) > touchSlop || abs(totalDy) > touchSlop)) {
-                    isDrag = true
-                    dragMode = if (abs(totalDx) > abs(totalDy)) 1 else if (startX < size.width * 0.5f) 2 else 3
-                    lastX = change.position.x
-                    lastY = change.position.y
-                }
-                if (isDrag) {
-                    val dx = change.position.x - lastX
-                    val dy = change.position.y - lastY
-                    lastX = change.position.x
-                    lastY = change.position.y
-                    change.consume()
-                    dispatchDragEvent(dragMode, dx / size.width.toFloat(), dy / size.height.toFloat(), callbacks)
-                }
-            }
-        }
-    }
-}
-
-private fun processTapRelease(
-    startX: Float,
-    width: Float,
-    lastTapTime: Long,
-    lastTapX: Float,
-    callbacks: PlayerGestureCallbacks,
-): Long {
-    val currentTime = System.currentTimeMillis()
-    val isDoubleTap = currentTime - lastTapTime < 300L && abs(startX - lastTapX) < 150f
-    if (isDoubleTap) {
-        dispatchDoubleTap(startX, width, callbacks)
-        return 0L
-    } else {
-        callbacks.onTap()
-        return currentTime
-    }
-}
-
-private fun dispatchDoubleTap(
-    startX: Float,
-    width: Float,
-    callbacks: PlayerGestureCallbacks,
-) {
-    if (startX < width * 0.4f) {
-        callbacks.onDoubleTapLeft()
-    } else if (startX > width * 0.6f) {
-        callbacks.onDoubleTapRight()
-    } else {
-        callbacks.onTap()
-    }
-}
-
-private fun dispatchDragEvent(
-    dragMode: Int,
-    ratioX: Float,
-    ratioY: Float,
-    callbacks: PlayerGestureCallbacks,
-) {
-    when (dragMode) {
-        1 -> callbacks.onHorizontalDrag(ratioX)
-        2 -> callbacks.onVerticalDragLeft(-ratioY)
-        3 -> callbacks.onVerticalDragRight(-ratioY)
-    }
-}
-
-private fun launchExternalPlayer(context: Context, video: VideoItem, packageName: String) {
-    val uri = extractUri(video) ?: return
-
-    val intent = Intent(Intent.ACTION_VIEW).apply {
-        if (isYouTubeUrl(uri.toString()) || isYouTubeUrl(video.name)) {
-            setDataAndType(uri, "text/html")
-        } else {
-            setDataAndType(uri, "video/*")
-        }
-        addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION or Intent.FLAG_ACTIVITY_NEW_TASK)
-        if (packageName.isNotEmpty()) {
-            setPackage(packageName)
-        }
-    }
-    runCatching {
-        context.startActivity(Intent.createChooser(intent, video.name))
-    }.onFailure {
-        runCatching {
-            context.startActivity(Intent(Intent.ACTION_VIEW, uri))
-        }
-    }
-}
-
-@Suppress("ReturnCount")
-private fun extractUri(video: VideoItem?): Uri? {
-    if (video == null) return null
-    val target = when {
-        !video.nativeUri.isNullOrEmpty() -> video.nativeUri
-        !video.url.isNullOrEmpty() -> video.url
-        video.path.isNotEmpty() -> return Uri.fromFile(File(video.path))
-        video.name.startsWith("http://") || video.name.startsWith("https://") || video.name.startsWith("content://") -> video.name
-        else -> null
-    } ?: return null
-    return runCatching { Uri.parse(target) }.getOrNull()
-}
-
-private fun isYouTubeUrl(urlStr: String?): Boolean {
-    if (urlStr == null) return false
-    val lower = urlStr.lowercase()
-    return lower.contains("youtube.com") || lower.contains("youtu.be")
-}
-
-private fun enterPipMode(activity: Activity?) {
-    if (activity != null && Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-        val params = PictureInPictureParams.Builder()
-            .setAspectRatio(Rational(16, 9))
-            .build()
-        activity.enterPictureInPictureMode(params)
-    }
-}
-
-@Composable
-private fun TopPlayerBar(
-    title: String,
-    aspectRatioMode: AspectRatioMode,
-    playbackSpeed: Float,
-    onBack: () -> Unit,
-    onOpenTracks: () -> Unit,
-    onCycleAspect: () -> Unit,
-    onCycleSpeed: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(16.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween,
-    ) {
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            modifier = Modifier.weight(1f),
-        ) {
-            IconButton(onClick = onBack) {
-                Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Retour", tint = White)
-            }
-            Spacer(modifier = Modifier.width(8.dp))
-            Text(
-                text = title,
-                color = White,
-                fontSize = 16.sp,
-                fontWeight = FontWeight.Bold,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-            )
-        }
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            IconButton(onClick = onOpenTracks) {
-                Icon(Icons.Filled.Subtitles, contentDescription = "Sous-titres & Audio", tint = White)
-            }
-            IconButton(onClick = onCycleAspect) {
-                Icon(Icons.Filled.AspectRatio, contentDescription = "Format : ${aspectRatioMode.label}", tint = White)
-            }
-            Button(
-                onClick = onCycleSpeed,
-                colors = ButtonDefaults.buttonColors(containerColor = Zinc800),
-                contentPadding = androidx.compose.foundation.layout.PaddingValues(horizontal = 8.dp, vertical = 4.dp),
-            ) {
-                Icon(Icons.Filled.Speed, contentDescription = null, tint = White, modifier = Modifier.size(16.dp))
-                Spacer(modifier = Modifier.width(4.dp))
-                Text("${playbackSpeed}x", color = White, fontSize = 12.sp)
-            }
-        }
-    }
-}
-
-@Composable
-private fun CenterPlayerControls(
-    isPlaying: Boolean,
-    hasNextVideo: Boolean,
-    onTogglePlay: () -> Unit,
-    onRewind: () -> Unit,
-    onForward: () -> Unit,
-    onNextVideo: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Row(
-        modifier = modifier,
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(24.dp),
-    ) {
-        IconButton(onClick = onRewind, modifier = Modifier.size(48.dp)) {
-            Icon(Icons.Filled.Replay10, contentDescription = "Reculer 10s", tint = White, modifier = Modifier.size(36.dp))
-        }
-
-        Box(
-            modifier = Modifier
-                .size(64.dp)
-                .clip(CircleShape)
-                .background(Red600)
-                .clickable(onClick = onTogglePlay),
-            contentAlignment = Alignment.Center,
-        ) {
-            Icon(
-                imageVector = if (isPlaying) Icons.Filled.Pause else Icons.Filled.PlayArrow,
-                contentDescription = if (isPlaying) "Pause" else "Lecture",
-                tint = White,
-                modifier = Modifier.size(40.dp),
-            )
-        }
-
-        IconButton(onClick = onForward, modifier = Modifier.size(48.dp)) {
-            Icon(Icons.Filled.Forward10, contentDescription = "Avancer 10s", tint = White, modifier = Modifier.size(36.dp))
-        }
-
-        if (hasNextVideo) {
-            IconButton(onClick = onNextVideo, modifier = Modifier.size(48.dp)) {
-                Icon(Icons.Filled.SkipNext, contentDescription = "Épisode suivant", tint = White, modifier = Modifier.size(36.dp))
-            }
-        }
-    }
-}
-
-@Composable
-private fun BottomPlayerBar(
-    positionMs: Long,
-    durationMs: Long,
-    isLocked: Boolean,
-    onSeek: (Long) -> Unit,
-    onToggleLock: () -> Unit,
-    onEnterPip: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    var isSeeking by remember { mutableStateOf(false) }
-    var seekPositionMs by remember { mutableLongStateOf(0L) }
-
-    val displayPos = if (isSeeking) seekPositionMs else positionMs
-
-    Column(
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(16.dp),
-    ) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceBetween,
-        ) {
-            Text(
-                text = "${formatTimeMs(displayPos)} / ${formatTimeMs(durationMs)}",
-                color = White,
-                fontSize = 12.sp,
-            )
-            Row {
-                IconButton(onClick = onEnterPip) {
-                    Icon(Icons.Filled.PictureInPicture, contentDescription = "PiP", tint = White)
-                }
-                IconButton(onClick = onToggleLock) {
-                    Icon(
-                        imageVector = if (isLocked) Icons.Filled.Lock else Icons.Filled.LockOpen,
-                        contentDescription = "Verrouiller",
-                        tint = White,
-                    )
-                }
-            }
-        }
-
-        val sliderValue = if (durationMs > 0L) {
-            displayPos.coerceIn(0L, durationMs).toFloat()
-        } else {
-            0f
-        }
-
-        Slider(
-            value = sliderValue,
-            onValueChange = {
-                if (durationMs > 0L) {
-                    isSeeking = true
-                    seekPositionMs = it.toLong()
-                }
-            },
-            onValueChangeFinished = {
-                if (durationMs > 0L) {
-                    onSeek(seekPositionMs)
-                    isSeeking = false
-                }
-            },
-            valueRange = 0f..durationMs.coerceAtLeast(1L).toFloat(),
-            colors = SliderDefaults.colors(
-                thumbColor = Red600,
-                activeTrackColor = Red600,
-                inactiveTrackColor = Zinc800,
-            ),
-            modifier = Modifier.fillMaxWidth(),
-        )
-    }
-}
-
-@Composable
-private fun GestureFeedbackCard(
-    feedback: GestureFeedback,
-    modifier: Modifier = Modifier,
-) {
-    Card(
-        shape = RoundedCornerShape(12.dp),
-        colors = CardDefaults.cardColors(containerColor = Zinc900.copy(alpha = 0.9f)),
-        modifier = modifier.padding(16.dp),
-    ) {
-        Row(
-            modifier = Modifier.padding(horizontal = 20.dp, vertical = 12.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(12.dp),
-        ) {
-            Icon(
-                imageVector = when (feedback.type) {
-                    FeedbackType.VOLUME -> Icons.Filled.VolumeUp
-                    FeedbackType.BRIGHTNESS -> Icons.Filled.Brightness6
-                    FeedbackType.SEEK_FORWARD -> Icons.Filled.Forward10
-                    FeedbackType.SEEK_REWIND -> Icons.Filled.Replay10
-                },
-                contentDescription = null,
-                tint = White,
-                modifier = Modifier.size(28.dp),
-            )
-            Text(text = feedback.text, color = White, fontSize = 14.sp, fontWeight = FontWeight.Bold)
-        }
-    }
-}
-
-@Composable
-private fun NextEpisodeOverlay(
-    nextVideo: VideoItem,
-    onPlayNext: () -> Unit,
-    onDismiss: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    var countdown by remember { mutableIntStateOf(5) }
-
-    LaunchedEffect(nextVideo) {
-        countdown = 5
-        while (countdown > 0) {
-            delay(1000L)
-            countdown--
-        }
-        onPlayNext()
-    }
-
-    Card(
-        shape = RoundedCornerShape(16.dp),
-        colors = CardDefaults.cardColors(containerColor = Zinc900),
-        modifier = modifier.fillMaxWidth(),
-    ) {
-        Column(modifier = Modifier.padding(16.dp)) {
-            Text("Épisode suivant dans $countdown s", color = Red600, fontSize = 12.sp, fontWeight = FontWeight.Bold)
-            Spacer(modifier = Modifier.height(4.dp))
-            Text(
-                text = nextVideo.cleanTitle ?: nextVideo.name,
-                color = White,
-                fontSize = 16.sp,
-                fontWeight = FontWeight.Bold,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-            )
-            Spacer(modifier = Modifier.height(12.dp))
-            Row(
-                horizontalArrangement = Arrangement.End,
-                modifier = Modifier.fillMaxWidth(),
-            ) {
-                Button(
-                    onClick = onDismiss,
-                    colors = ButtonDefaults.buttonColors(containerColor = Zinc800),
-                ) {
-                    Text("Fermer", color = White)
-                }
-                Spacer(modifier = Modifier.width(8.dp))
-                Button(
-                    onClick = onPlayNext,
-                    colors = ButtonDefaults.buttonColors(containerColor = Red600),
-                ) {
-                    Text("Lire maintenant", color = White)
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun TracksSelectionSheet(
-    audioTracks: List<AudioTrackUiState>,
-    subtitleTracks: List<SubtitleTrackUiState>,
-    onSelectAudio: (String) -> Unit,
-    onSelectSubtitle: (String?) -> Unit,
-    onPickSubtitleFile: () -> Unit,
-    onDismiss: () -> Unit,
-) {
-    val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
-
-    ModalBottomSheet(
-        onDismissRequest = onDismiss,
-        sheetState = sheetState,
-        containerColor = Zinc900,
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(20.dp),
-        ) {
-            Text("Pistes Audio", color = White, fontSize = 18.sp, fontWeight = FontWeight.Bold)
-            Spacer(modifier = Modifier.height(8.dp))
-            if (audioTracks.isEmpty()) {
-                Text("Aucune piste audio détectée", color = Color.Gray, fontSize = 14.sp)
-            } else {
-                audioTracks.forEach { track ->
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically,
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .clickable { onSelectAudio(track.id) }
-                            .padding(vertical = 6.dp),
-                    ) {
-                        RadioButton(
-                            selected = track.isSelected,
-                            onClick = { onSelectAudio(track.id) },
-                            colors = RadioButtonDefaults.colors(selectedColor = Red600),
-                        )
-                        Text(track.label, color = White, fontSize = 14.sp)
-                    }
-                }
-            }
-
-            Spacer(modifier = Modifier.height(20.dp))
-            Text("Sous-titres", color = White, fontSize = 18.sp, fontWeight = FontWeight.Bold)
-            Spacer(modifier = Modifier.height(8.dp))
-
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .clickable { onSelectSubtitle(null) }
-                    .padding(vertical = 6.dp),
-            ) {
-                RadioButton(
-                    selected = subtitleTracks.none { it.isSelected },
-                    onClick = { onSelectSubtitle(null) },
-                    colors = RadioButtonDefaults.colors(selectedColor = Red600),
-                )
-                Text("Désactivés", color = White, fontSize = 14.sp)
-            }
-
-            subtitleTracks.forEach { track ->
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .clickable { onSelectSubtitle(track.id) }
-                        .padding(vertical = 6.dp),
-                ) {
-                    RadioButton(
-                        selected = track.isSelected,
-                        onClick = { onSelectSubtitle(track.id) },
-                        colors = RadioButtonDefaults.colors(selectedColor = Red600),
-                    )
-                    Text(
-                        text = if (track.isExternal) "${track.label} (Fichier externe)" else track.label,
-                        color = White,
-                        fontSize = 14.sp,
-                    )
-                }
-            }
-
-            Spacer(modifier = Modifier.height(12.dp))
-            Button(
-                onClick = onPickSubtitleFile,
-                colors = ButtonDefaults.buttonColors(containerColor = Zinc800),
-                modifier = Modifier.fillMaxWidth(),
-            ) {
-                Icon(Icons.Filled.Add, contentDescription = null, tint = White)
-                Spacer(modifier = Modifier.width(8.dp))
-                Text("Importer un fichier .srt", color = White)
-            }
-            Spacer(modifier = Modifier.height(20.dp))
-        }
-    }
-}
-
-private fun formatTimeMs(ms: Long): String {
-    val totalSeconds = (ms / 1000).coerceAtLeast(0)
-    val hours = totalSeconds / 3600
-    val minutes = (totalSeconds % 3600) / 60
-    val seconds = totalSeconds % 60
-    return if (hours > 0) {
-        String.format(Locale.US, "%d:%02d:%02d", hours, minutes, seconds)
-    } else {
-        String.format(Locale.US, "%02d:%02d", minutes, seconds)
-    }
-}
-
-
-@Suppress("UnusedPrivateMember", "LongMethod")
-@SuppressLint("SetJavaScriptEnabled")
-@Composable
-private fun YouTubePlayerView(
-    videoId: String,
-    initialPositionMs: Long,
-    isPlaying: Boolean,
-    positionMs: Long,
-    onPositionChanged: (positionMs: Long, durationMs: Long) -> Unit,
-    onPlayingStateChanged: (isPlaying: Boolean) -> Unit,
-    onEnded: () -> Unit,
-    onError: (error: String?) -> Unit,
-    onBuffering: (isBuffering: Boolean) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    var webViewRef by remember { mutableStateOf<WebView?>(null) }
-    var isJsReady by remember { mutableStateOf(false) }
-    var lastWebPositionMs by remember { mutableLongStateOf(initialPositionMs) }
-
-    val startSeconds = (initialPositionMs / 1000).coerceAtLeast(0)
-
-    val htmlContent = remember(videoId, startSeconds) {
-        """
-        <!DOCTYPE html>
-        <html>
-        <head>
-            <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no">
-            <style>
-                body, html { margin: 0; padding: 0; width: 100%; height: 100%; background-color: #000000; overflow: hidden; }
-                #player { width: 100%; height: 100%; }
-            </style>
-        </head>
-        <body>
-            <div id="player"></div>
-            <script>
-                var tag = document.createElement('script');
-                tag.src = "https://www.youtube.com/iframe_api";
-                var firstScriptTag = document.getElementsByTagName('script')[0];
-                firstScriptTag.parentNode.insertBefore(tag, firstScriptTag);
-
-                var player;
-                function onYouTubeIframeAPIReady() {
-                    player = new YT.Player('player', {
-                        height: '100%',
-                        width: '100%',
-                        videoId: '$videoId',
-                        playerVars: {
-                            'autoplay': 1,
-                            'controls': 0,
-                            'rel': 0,
-                            'fs': 0,
-                            'playsinline': 1,
-                            'start': $startSeconds,
-                            'enablejsapi': 1,
-                            'modestbranding': 1
-                        },
-                        events: {
-                            'onReady': onPlayerReady,
-                            'onStateChange': onPlayerStateChange,
-                            'onError': onPlayerError
-                        }
-                    });
-                }
-                function onPlayerReady(event) {
-                    if (window.AndroidBridge) {
-                        window.AndroidBridge.onPlayerReady();
-                    }
-                    setInterval(function() {
-                        if (player && player.getPlayerState && player.getPlayerState() === 1) {
-                            var curTime = player.getCurrentTime() || 0;
-                            var dur = player.getDuration() || 0;
-                            if (window.AndroidBridge) {
-                                window.AndroidBridge.onProgress(curTime, dur);
-                            }
-                        }
-                    }, 500);
-                }
-                function onPlayerStateChange(event) {
-                    if (window.AndroidBridge) {
-                        var curTime = player && player.getCurrentTime ? player.getCurrentTime() : 0;
-                        var dur = player && player.getDuration ? player.getDuration() : 0;
-                        window.AndroidBridge.onStateChange(event.data, curTime, dur);
-                    }
-                }
-                function onPlayerError(err) {
-                    if (window.AndroidBridge) {
-                        window.AndroidBridge.onError("" + err.data);
-                    }
-                }
-                function playVideo() { if(player && player.playVideo) player.playVideo(); }
-                function pauseVideo() { if(player && player.pauseVideo) player.pauseVideo(); }
-                function seekTo(sec) { if(player && player.seekTo) player.seekTo(sec, true); }
-            </script>
-        </body>
-        </html>
-        """.trimIndent()
-    }
-
-    DisposableEffect(videoId) {
-        onDispose {
-            webViewRef?.destroy()
-        }
-    }
-
-    LaunchedEffect(isPlaying, isJsReady) {
-        if (isJsReady) {
-            if (isPlaying) {
-                webViewRef?.evaluateJavascript("playVideo()", null)
-            } else {
-                webViewRef?.evaluateJavascript("pauseVideo()", null)
-            }
-        }
-    }
-
-    LaunchedEffect(positionMs, isJsReady) {
-        if (isJsReady && kotlin.math.abs(positionMs - lastWebPositionMs) > 1500L) {
-            val sec = positionMs / 1000f
-            lastWebPositionMs = positionMs
-            webViewRef?.evaluateJavascript("seekTo($sec)", null)
-        }
-    }
-
-    AndroidView(
-        factory = { ctx ->
-            WebView(ctx).apply {
-                layoutParams = ViewGroup.LayoutParams(
-                    ViewGroup.LayoutParams.MATCH_PARENT,
-                    ViewGroup.LayoutParams.MATCH_PARENT
-                )
-                settings.javaScriptEnabled = true
-                settings.domStorageEnabled = true
-                settings.mediaPlaybackRequiresUserGesture = false
-                webViewClient = object : WebViewClient() {}
-                webChromeClient = object : WebChromeClient() {}
-                val mainScope = CoroutineScope(Dispatchers.Main)
-                addJavascriptInterface(
-                    YouTubeBridge(
-                        mainScope = mainScope,
-                        onReady = { isJsReady = true },
-                        onStateChange = { state, currentTimeSec, durationSec ->
-                            when (state) {
-                                1 -> { // PLAYING
-                                    onBuffering(false)
-                                    onError(null)
-                                    onPlayingStateChanged(true)
-                                    val pos = (currentTimeSec * 1000).toLong()
-                                    lastWebPositionMs = pos
-                                    onPositionChanged(pos, (durationSec * 1000).toLong())
-                                }
-                                2 -> { // PAUSED
-                                    onBuffering(false)
-                                    onPlayingStateChanged(false)
-                                    val pos = (currentTimeSec * 1000).toLong()
-                                    lastWebPositionMs = pos
-                                    onPositionChanged(pos, (durationSec * 1000).toLong())
-                                }
-                                3 -> { // BUFFERING
-                                    onBuffering(true)
-                                }
-                                0 -> { // ENDED
-                                    onBuffering(false)
-                                    onPlayingStateChanged(false)
-                                    onEnded()
-                                }
-                                else -> Unit
-                            }
-                        },
-                        onProgress = { currentTimeSec, durationSec ->
-                            val pos = (currentTimeSec * 1000).toLong()
-                            lastWebPositionMs = pos
-                            onPositionChanged(pos, (durationSec * 1000).toLong())
-                        },
-                        onError = { errorCode ->
-                            onBuffering(false)
-                            onError("Erreur de lecture YouTube (code $errorCode)")
-                        },
-                    ),
-                    "AndroidBridge",
-                )
-
-                loadDataWithBaseURL("https://www.youtube.com", htmlContent, "text/html", "UTF-8", null)
-                webViewRef = this
-            }
-        },
-        modifier = modifier.fillMaxSize()
-    )
-}
-
-private class YouTubeBridge(
-    private val mainScope: CoroutineScope,
-    private val onReady: () -> Unit,
-    private val onStateChange: (state: Int, currentTimeSec: Float, durationSec: Float) -> Unit,
-    private val onProgress: (currentTimeSec: Float, durationSec: Float) -> Unit,
-    private val onError: (errorCode: String) -> Unit,
-) {
-    @JavascriptInterface
-    fun onPlayerReady() {
-        mainScope.launch { onReady() }
-    }
-
-    @JavascriptInterface
-    fun onStateChange(state: Int, currentTimeSec: Float, durationSec: Float) {
-        mainScope.launch { onStateChange(state, currentTimeSec, durationSec) }
-    }
-
-    @JavascriptInterface
-    fun onProgress(currentTimeSec: Float, durationSec: Float) {
-        mainScope.launch { onProgress(currentTimeSec, durationSec) }
-    }
-
-    @JavascriptInterface
-    fun onError(errorCode: String) {
-        mainScope.launch { onError(errorCode) }
-    }
-}
